pyproject.toml
[project]
name = "strmlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.50.0",
    "python-dotenv>=1.0.1",
]
